Fix localstorage error when server rendering and add loading states.
localstorage is not defined when we render on the server. We wrap up
localstorage into a createStorageValue helper with a browser check.

As we don't have access to localstorage when we render on the server,
we need to render some skeleton loading states. We add some extra loading
state to the animation context and use this to display skeleton components